requirements/base.txt
PySide6>=6.6
PyYAML>=6.0
numpy>=1.26
Pillow>=10.0
